Refactored code: Pull-up var/method, Extract class, Polymorphism
