+++ src/main/java/com/networknt/schema/AnnotationsHandler.java
@@ -1,0 +1,28 @@
+package com.networknt.schema;
+import com.fasterxml.jackson.databind.JsonNode;
+
+abstract class AnnotationValueHandler {
+    public abstract Object getAnnotationValue(JsonNode schemaNode);
+}
+
+class TextualValueHandler extends AnnotationValueHandler {
+    @Override
+    public Object getAnnotationValue(JsonNode schemaNode) {
+        return schemaNode.textValue();
+    }
+}
+
+class NumericValueHandler extends AnnotationValueHandler {
+    @Override
+    public Object getAnnotationValue(JsonNode schemaNode) {
+        return schemaNode.numberValue();
+    }
+}
+
+class ObjectValueHandler extends AnnotationValueHandler {
+    @Override
+    public Object getAnnotationValue(JsonNode schemaNode) {
+        return schemaNode;
+    }
+} 
+
